user get otr not

+++ src/app/modules/user/user.service.ts
@@ -1,5 +1,5 @@
 import { UserModel } from "./user.model";
-import User from "./user.interface";
+import {User} from "./user.interface";
 
 const createUserIntoDb = async (user: User) => {
   const result = await UserModel.create(user);
@@ -17,32 +17,94 @@ const createUserIntoDb = async (user: User) => {
 };
 
 const getUsersFromDb = async () => {
-  const result = await UserModel.find().select("-_id username fullName age email address");
+  const result = await UserModel.find().select(
+    "-_id username fullName age email address"
+  );
 
   return result;
 };
 
 const getUserFromDb = async (userId: number) => {
-  const result = await UserModel.findOne({ userId }).select("-_id username fullName age email address");
-
-  return result;
+  if (await UserModel.isUserExist(Number(userId))) {
+    const result = await UserModel.findOne({ userId }).select(
+      "-_id username fullName age email address"
+    );
+    return {
+      success: true,
+      message: "Users fetched successfully!",
+      data: result,
+    };
+  } else {
+    return {
+      success: false,
+      message: "User not found",
+      error: {
+        code: 404,
+        description: "User not found!",
+      },
+    };
+  }
 };
 
 const deleteUserFromDb = async (userId: number) => {
-  const result = await UserModel.deleteOne({ userId });
+  
 
-  return result;
+  if (await UserModel.isUserExist(Number(userId))) {
+    const result = await UserModel.deleteOne({ userId });
+    return {
+      success: true,
+      message: "User deleted successfully!",
+      data: null,
+    };
+  } else {
+    return {
+      success: false,
+      message: "User not found",
+      error: {
+        code: 404,
+        description: "User not found!",
+      },
+    };
+  }
+
+  
 };
 
-const updateUserIntoDb = async (user: User, userId: number) => {
-  const result = await UserModel.updateOne(
-    { userId },
-    {
-      $set: user,
-    }
-  );
 
-  return result;
+
+
+  
+    
+    const updateUserIntoDb = async (user: User, userId: number) => {
+      if (await UserModel.isUserExist(Number(userId))) {
+      const result = await UserModel.updateOne(
+        { userId },
+        {
+          $set: user,
+        }
+      );
+
+      const newUser =  await UserModel.findOne({ userId }).select(
+        "-_id username fullName age email address"
+      );
+    return{
+      
+    "success": true,
+    "message": "User updated successfully!",
+    "data": newUser
+    }
+  } 
+  else {
+    return {
+      success: false,
+      message: "User not found",
+      error: {
+        code: 404,
+        description: "User not found!",
+      },
+    };
+  }
+
 };
 
 const addOrderIntoDb = async (order: any, userId: number) => {
@@ -62,7 +124,6 @@ const addOrderIntoDb = async (order: any, userId: number) => {
   } else return {};
 };
 
-
 const getOrderFromDb = async (userId: number) => {
   const result = await UserModel.findOne({ userId }).select("-_id orders");
 
@@ -71,10 +132,14 @@ const getOrderFromDb = async (userId: number) => {
 
 const getOrderTotalFromDb = async (userId: number) => {
   const result = await UserModel.findOne({ userId }).select("-_id orders");
+  let sum = 0;
+  if (result?.orders) {
+    result?.orders?.forEach((el) => {
+      sum = sum + el.quantity * el.price;
+    });
+  }
 
-  console.log(result)
-
-  return result;
+  return sum;
 };
 
 export const userService = {
@@ -85,5 +150,5 @@ export const userService = {
   updateUserIntoDb,
   addOrderIntoDb,
   getOrderFromDb,
-  getOrderTotalFromDb
+  getOrderTotalFromDb,
 };
